requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI & LLM Framework
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# OpenAI Integration
openai>=1.0.0

# Web Search Tools
tavily-python>=0.3.0

# Additional utilities that might be needed
pydantic>=2.0.0
requests>=2.31.0
typing-extensions>=4.7.0

# Optional: For better performance
aiohttp>=3.8.0
asyncio

# Development tools (optional - comment out for production)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
